fix z index bug with login
