+++ src/pages/Welcome/Welcome.tsx
@@ -4,8 +4,11 @@ import WelcomeImg from "../../assets/welcome_placeholder.png";
 import WelcomeJournal from "../../assets/welcome_journal.svg";
 import WelcomeTrack from "../../assets/welcome_track.svg";
 import WelcomeShare from "../../assets/welcome_share.svg";
-import SignIn from "../SignIn/SignIn";
 import Column from "../../components/layout/Column";
+import Button from "../../components/Button";
+import ForgotPasswordModal from "./ForgotPasswordModal";
+import SignUpModal from "./SignUpModal";
+import LoginModal from "./LoginModal";
 
 function FeatureCard({
   title,
@@ -48,12 +51,30 @@ export default function Welcome() {
     },
   ];
 
+  const openLogin = () => {
+    window.HSOverlay.open(document.querySelector("#login-modal"));
+  };
+
+  const openSignUp = () => {
+    window.HSOverlay.open(document.querySelector("#sign-up-modal"));
+  };
+
+  const openForgotPassword = () => {
+    window.HSOverlay.open(document.querySelector("#forgot-pass-modal"));
+  };
+
   return (
     <div>
       {/* Navbar */}
       <Row className="sticky top-0 z-10 justify-between rounded-b-2xl bg-light_yellow pb-5 pl-12 pr-12 pt-8 md:pl-28">
         <img src={HappinessAppText} className="pr-4" />
-        <SignIn signUp={false} />
+        <Button
+          label="Log In"
+          associatedModalId="login-modal"
+          onClick={openLogin}
+          variation="OUTLINED"
+          classNameText="font-bold"
+        />
       </Row>
 
       {/* Fixed Div */}
@@ -63,7 +84,14 @@ export default function Welcome() {
             <p className="mb-12 text-5xl font-extrabold text-secondary">
               Begin Your <br /> Happiness Journey.
             </p>
-            <SignIn signUp={true} />
+            <Button
+              label="Get Started"
+              associatedModalId="sign-in-modal"
+              onClick={openSignUp}
+              variation="FILLED"
+              classNameBtn="h-auto bg-yellow px-7 py-5"
+              classNameText="text-[22px]"
+            />
           </Column>
           <img
             src={WelcomeImg}
@@ -113,6 +141,13 @@ export default function Welcome() {
           </Column>
         </Row>
       </div>
+      <ForgotPasswordModal id="forgot-pass-modal" onLoginClick={openLogin} />
+      <SignUpModal id="sign-up-modal" onLoginClick={openLogin} />
+      <LoginModal
+        id="login-modal"
+        onCreateAccountClick={openSignUp}
+        onForgotPassword={openForgotPassword}
+      />
     </div>
   );
 }
